+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 2.4.2
 - Fix Forgotten Village Terminators
+- Fix Pumpkin Burrow not releasing Warden on break
 
 # 2.4.1
 - Update Chinese (Simplified) Translations (zh_cn) (Credits: Crown-Fairy) for Village Update